% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\name{alpha.drop}
\alias{alpha.drop}
\title{Produce Cronbach alpha if dropped output using quosures}
\usage{
alpha.drop(data, items, name = NULL, neg.corr = FALSE)
}
\arguments{
\item{data}{data frame}

\item{items}{quosure of items to analyze}

\item{name}{optional vector of item names/wording}

\item{neg.corr}{are some items negatively correlated? This reverses those items to get an accurate estimate of alpha}
}
\value{
cronbach alpha output from the psych package's 'alpha' function
}
\description{
This function reports the Cronbach's alpha for a group of items if each item were dropped.
}
