% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\name{alpha_drop}
\alias{alpha_drop}
\title{Produce Cronbach alpha output using quosures}
\usage{
alpha_drop(data, quosure, name = NULL, neg.corr = FALSE)
}
\arguments{
\item{data}{data frame}

\item{neg.corr}{are some items negatively correlated? This reverses those items to get an accurate estimate of alpha}

\item{vector}{quosure of items to analyze}
}
\value{
cronbach alpha output from the psych package's 'alpha' function
}
\description{
This function produces info about Cronbach's alpha for a group of variables
}
