% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidying.r
\name{bigO.drop}
\alias{bigO.drop}
\title{Ben Lira's function for assessing and dropping multivariate outliers. T}
\usage{
bigO.drop(data, vars = NULL, exclude = NULL, p = 0.001)
}
\arguments{
\item{data}{data frame}

\item{vars}{vector of subset of data to analyze. Takes either varnames or a numeric vector of column numbers.}

\item{exclude}{vector of subset of data to exclude from analysis. Takes either varnames or a numeric vector of column numbers.}

\item{p}{level at which to test the significance of outliers, by testing the Chi2 parameter of the alpha distribution.}
}
\value{
cases excluded, a dataset of remaining cases, and a graph showing excluded vs. retained cases.
}
\description{
The function provides information on multivariate outliers, but also produces
a dataset with multivariate outliers excluded, which can be saved to a new object, or used to overwrite the data set.
This calculates Mahalanobis distance for each observation and calculates multivariate outliers using Chi2 testing.
}
