% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{candyhierarchy_localities}
\alias{candyhierarchy_localities}
\title{Character vector of 2460 localities}
\format{A character vector of localities from The 2017 Halloween Candy Hierarchy survey
\describe{
  \item{candyhierarchy_localities}{Not clean locality data}
}}
\usage{
candyhierarchy_localities
}
\description{
A character vector of "Q5: STATE, PROVINCE, COUNTY, ETC" column 
of locality data collected from 
The 2017 Halloween Candy Hierarchy survey. Url can be found in details.
}
\details{
https://boingboing.net/2017/10/30/the-2017-halloween-candy-hiera.html
}
\keyword{datasets}
