% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.r
\name{corr_table}
\alias{corr_table}
\title{Basic correlation table}
\usage{
corr_table(df, vars, rows = auto, cols = auto, copy = FALSE)
}
\arguments{
\item{df}{data frame}

\item{vars}{quosure of variables to analyze}
}
\description{
Basic correlation table
}
