% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.r
\name{corrtrol}
\alias{corrtrol}
\title{Basic correlation table}
\usage{
corrtrol(
  df,
  control.vars,
  rows = auto,
  cols = auto,
  triangle = "lower",
  copy = FALSE
)
}
\arguments{
\item{df}{data frame of vars to analyse}

\item{control.vars}{variables to control for in analysis}

\item{rows}{number of variables used as rows; used in conjunction with cols. Default is all vars}

\item{cols}{number of variables used as columns; used in conjunction with rowss. Default is all vars}

\item{triangle}{toggle lower/upper triangle output of correlations}

\item{copy}{Should this table be 'copiable' or formatted?}
}
\description{
Basic correlation table
}
