% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_creation.r
\name{cv_bake}
\alias{cv_bake}
\title{Generate a composite variable}
\usage{
cv_bake(data, new, ingred.list)
}
\arguments{
\item{data}{data frame}

\item{new}{name of new variable to create}

\item{ingred.list}{quosure of variables to average into composite}
}
\value{
a new variable composite of individual items
}
\description{
This function creates a composite variable of a group of individual items in your data frame.
}
