% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigating.r
\name{factorize}
\alias{factorize}
\title{Perform factor analysis and get useful output using quosures}
\usage{
factorize(data, vector, factors = 1, rotation = "varimax")
}
\arguments{
\item{data}{data frame}

\item{vector}{quosure of items to analyze}

\item{factors}{the number of factors to fit}

\item{rotation}{Type of rotation; drawing from rotation options in the 'psych' package's factanal function.}
}
\value{
factor analysis output and scree plot
}
\description{
This function creates a composite variable of a group of individual items in your data frame.
}
