% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{fingerprint}
\alias{fingerprint}
\title{Find fingerprint of string}
\usage{
fingerprint(s)
}
\arguments{
\item{s}{chacacter vector}
}
\value{
fingerprint of given vector in a character vector of same size
}
\description{
Extremely simple fingerprint function for takeoff
}
\examples{
school <- c("school","School","school","School","School ","SCHOOL","School")
fingerprint(school)
}
