% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feedback.R
\name{get_feedback}
\alias{get_feedback}
\title{Get feedback from user whether to replace a value}
\usage{
get_feedback(x = "NoT CleAn ", replacement = "Clean", ask_user = T)
}
\arguments{
\item{x}{One-lengt charachter}

\item{replacement}{One-lengt charachter proposed as a replacement to x}

\item{ask_user}{Logical value in order to ask or accept right away replacement value}
}
\value{
Final value that will be used instead of x
}
\description{
Get feedback from user whether to replace provided one lengt character
}
\examples{
get_feedback(x = "School ", replacement = "school", ask_user = FALSE)
\dontrun{
get_feedback(x = "school", replacement = "house")
}
}
