% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.r
\name{HARcorr}
\alias{HARcorr}
\title{Generate a formatted, APA-ready correlation table with descriptives}
\usage{
HARcorr(
  df,
  vars,
  describe = TRUE,
  numbers = TRUE,
  headers = NULL,
  spots = NULL,
  copy = FALSE,
  names = NULL,
  full.labels = FALSE
)
}
\arguments{
\item{df}{data frame}

\item{vars}{quosure of variables to analyze}

\item{describe}{allows you to toggle on/off the descriptives (mean, standard deviation, and n) at the bottom of the table by specifying describe = TRUE/FALSE.}

\item{numbers}{allows you to toggle on/off the numbers used to reference variables on the horizontal axis by specifying numbers = TRUE/FALSE. If numbers = FALSE, variable names are used as the horizontal axis reference.}

\item{headers}{The options headers and spots are used together. They each take a vector as an argument. Headers specifies subheaders for groupings of variables; headers takes a character vector.}

\item{spots}{spots specifies the row before which you would like to insert the headers. Spots takes a numeric vector.}

\item{copy}{Copy allows you to toggle on/off the table formatting. If copy = FALSE, kable() formatting is applied to make the table formatted nicely for easy of use and presentation. If copy = TRUE, the formatting is removed and HARcorr() returns a data table, which makes the values more easily copied into other programs such as Excel or Word.}

\item{names}{a character vector containing full names of variables. These replace variable names to increase clarity.}

\item{full.labels}{Would you also like names on the top axis?}
}
\value{
a formatted correlation table
}
\description{
Below is code for producing formatted, paper-ready correlation tables with 'HARcorr' (i.e. hot-and-ready correlations; I like fun function names, sue me).
I tried to write the code to minimize the amount of packages the function required, and to use packages that you probably already use frequently.
The code produces bivariate correlation tables complete with many features that align the output with APA style and with standards and information you might
want for publishing your research. The main features include:
Bivariate correlations displayed on the lower half of the correlation matrix (omitted for the upper half), with 1's on the diagonal omitted.
Stars to indicate significance level, with a dagger indicating marginal significance (p < .10)
leading 0's removed from correlations; correlations rounded to 2 decimal places
Numbered variable names vertically, along with corresponding numbers horizontally.
Descriptive statistics for each variable (mean, standard deviation, n) displayed for each variable at the bottom of the matrix.
Per APA, numbers greater than 1,000 have commas (1,000 vs. 1000), and means and standard deviations between \link{-1, 1} that have a maximum/minimum
of greater than 1/-1 have a leading 0 (e.g. 0.57). "Means" of binary variables are presented as percentages and SD for these variables are omitted.
The options to include headers for groups of variables (e.g. including a header "Covariates" before rows of covariate variables)
For me, the inclusion of APA formatting, numbered variables, and descriptives are a huge improvement over previous correlation table functions I've used,
and save me a lot of time in formatting tables in papers.
}
