% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\name{histomize}
\alias{histomize}
\title{Produces formatted histogram}
\usage{
histomize(data, x, type = "factor", title, subtitle, xlab, ylab = "Count")
}
\arguments{
\item{data}{data frame}

\item{x}{variable to visualize}

\item{type}{specifies whether the variable is a factor or numeric variable. Because numeric variables have many thin bins, this removes proportion
and count data that would clog the graph}

\item{title}{main title for the top of the graph}

\item{subtitle}{subtitle to go under the main title}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis; default is "Count"}
}
\value{
Formatted histogram
}
\description{
This function, exported from the local package 'vizzify', produces a histogram with count and proportions for each bin.
}
