% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigating.r
\name{STATAtab}
\alias{STATAtab}
\title{Produces a table of frequency information for a variable}
\usage{
STATAtab(data, ...)
}
\arguments{
\item{data}{data frame}

\item{...}{variable to analyze}
}
\value{
a table with frequency information
}
\description{
This function produces a table of frequency information for a variable, including the n and \% for each level, as well as cumulative n and \%.
This function is designed to mirror the comprehensive frequency output provided by STATA.
}
