% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\name{table.alpha}
\alias{table.alpha}
\title{Produce Cronbach alpha output using quosures}
\usage{
table.alpha(data, x)
}
\arguments{
\item{data}{data frame}

\item{x}{quosure of composites to analyze}

\item{neg.corr}{are some items negatively correlated? This reverses those items to get an accurate estimate of alpha}
}
\value{
a vector of Cronbach's alphas for a list of composite variables
}
\description{
This function produces a vector of Cronbach's alphas for a list of composite variables.
Note that this function requires quosures including the consituent items of each composite
to exist and have already been created. These are necessary to assess alpha for each composite.
}
