% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigating.r
\name{table.freq}
\alias{table.freq}
\title{Produce a table of frequencies for a group of variables using quosures}
\usage{
table.freq(data, vars)
}
\arguments{
\item{data}{data frame}

\item{vars}{quosure of items to analyze}
}
\value{
a data.frame of frequencies for analyzed variables
}
\description{
This function produces a table of frequencies for each of variables in a group
}
