% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\name{tabshell}
\alias{tabshell}
\title{Produce a formatted table of proportions of frequency for a single variable}
\usage{
tabshell(data, ..., wide = TRUE)
}
\arguments{
\item{data}{data frame}

\item{...}{name of single variabe}

\item{wide}{should the table be full width or compressed?}
}
\value{
a formatted table of proportions of frequency for a single variable
}
\description{
Produce a formatted table of proportions of frequency for a single variable
}
