% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigating.r
\name{alphatize}
\alias{alphatize}
\title{Produce Cronbach alpha output using quosures}
\usage{
alphatize(data, vector, neg.corr = FALSE)
}
\arguments{
\item{data}{data frame}

\item{vector}{quosure of items to analyze}

\item{neg.corr}{are some items negatively correlated? This reverses those items to get an accurate estimate of alpha}
}
\value{
cronbach alpha output from the psych package's 'alpha' function
}
\description{
This function produces info about Cronbach's alpha for a group of variables
}
