% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean character vector}
\usage{
clean(s, ask_user = TRUE)
}
\arguments{
\item{s}{chacacter vector}

\item{ask_user}{Logical value in order to ask user her opinoin or automatically clean}
}
\value{
chacacter vector of clean/unified strings
}
\description{
Clean character vector automatically with user feedback
}
