% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\name{nutshell}
\alias{nutshell}
\title{Produce a short report of descriptives}
\usage{
nutshell(data, vector, type = "family")
}
\arguments{
\item{data}{data frame}

\item{vector}{quosure of items to analyze}

\item{type}{is this for a single variable ('lonely') or a group of variables ('family')?}
}
\value{
a table of n, mean, and sd for variables
}
\description{
This function produces descriptives in a 'nutshell'
}
